+++ d2/src/main/java/com/linkedin/d2/xds/balancer/XdsFsTogglingLoadBalancerFactory.java
@@ -81,6 +81,7 @@ public class XdsFsTogglingLoadBalancerFactory
   private final CanaryDistributionProvider _canaryDistributionProvider;
   private final FailoutConfigProviderFactory _failoutConfigProviderFactory;
   private final boolean _loadBalanceStreamException;
+  private final boolean _enablePotentialClientsCache;
 
   @Deprecated
   public XdsFsTogglingLoadBalancerFactory(long timeout, TimeUnit timeoutUnit, String fsBasePath,
@@ -98,6 +99,7 @@ public class XdsFsTogglingLoadBalancerFactory
         canaryDistributionProvider, false);
   }
 
+  @Deprecated
   public XdsFsTogglingLoadBalancerFactory(long timeout, TimeUnit timeoutUnit, String fsBasePath,
       Map<String, TransportClientFactory> clientFactories,
       Map<String, LoadBalancerStrategyFactory<? extends LoadBalancerStrategy>> loadBalancerStrategyFactories,
@@ -107,6 +109,22 @@ public class XdsFsTogglingLoadBalancerFactory
       DeterministicSubsettingMetadataProvider deterministicSubsettingMetadataProvider,
       FailoutConfigProviderFactory failoutConfigProviderFactory, CanaryDistributionProvider canaryDistributionProvider,
       boolean loadBalanceStreamException)
+  {
+    this(timeout, timeoutUnit, fsBasePath, clientFactories, loadBalancerStrategyFactories, d2ServicePath, sslContext,
+        sslParameters, isSSLEnabled, clientServicesConfig, partitionAccessorRegistry, sslSessionValidatorFactory,
+        d2ClientJmxManager, deterministicSubsettingMetadataProvider, failoutConfigProviderFactory,
+        canaryDistributionProvider, loadBalanceStreamException, false);
+  }
+
+  public XdsFsTogglingLoadBalancerFactory(long timeout, TimeUnit timeoutUnit, String fsBasePath,
+      Map<String, TransportClientFactory> clientFactories,
+      Map<String, LoadBalancerStrategyFactory<? extends LoadBalancerStrategy>> loadBalancerStrategyFactories,
+      String d2ServicePath, SSLContext sslContext, SSLParameters sslParameters, boolean isSSLEnabled,
+      Map<String, Map<String, Object>> clientServicesConfig, PartitionAccessorRegistry partitionAccessorRegistry,
+      SslSessionValidatorFactory sslSessionValidatorFactory, D2ClientJmxManager d2ClientJmxManager,
+      DeterministicSubsettingMetadataProvider deterministicSubsettingMetadataProvider,
+      FailoutConfigProviderFactory failoutConfigProviderFactory, CanaryDistributionProvider canaryDistributionProvider,
+      boolean loadBalanceStreamException, boolean enablePotentialClientsCache)
   {
     _lbTimeout = timeout;
     _lbTimeoutUnit = timeoutUnit;
@@ -125,6 +143,7 @@ public class XdsFsTogglingLoadBalancerFactory
     _failoutConfigProviderFactory = failoutConfigProviderFactory;
     _canaryDistributionProvider = canaryDistributionProvider;
     _loadBalanceStreamException = loadBalanceStreamException;
+    _enablePotentialClientsCache = enablePotentialClientsCache;
   }
 
   public TogglingLoadBalancer create(ScheduledExecutorService executorService, XdsToD2PropertiesAdaptor xdsAdaptor)
@@ -166,7 +185,7 @@ public class XdsFsTogglingLoadBalancerFactory
         new SimpleLoadBalancerState(executorService, uriBus, clusterBus, serviceBus, _clientFactories,
             _loadBalancerStrategyFactories, _sslContext, _sslParameters, _isSSLEnabled, _partitionAccessorRegistry,
             _sslSessionValidatorFactory, _deterministicSubsettingMetadataProvider, _canaryDistributionProvider,
-            _loadBalanceStreamException);
+            _loadBalanceStreamException, _enablePotentialClientsCache);
     _d2ClientJmxManager.setSimpleLoadBalancerState(state);
 
     SimpleLoadBalancer balancer =